Pool each line type separately
