+++ osu.Game.Rulesets.Sentakki/UI/Components/HitObjectLine/LineType.cs
@@ -1,0 +1,11 @@
+namespace osu.Game.Rulesets.Sentakki.UI.Components.HitObjectLine
+{
+    public enum LineType
+    {
+        Single,
+        OneAway,
+        TwoAway,
+        ThreeAway,
+        FullCircle,
+    }
+}
